+++ src/build.rs
@@ -2,7 +2,6 @@ use crate::cmd::{Command, MountKind, Runnable, SandboxBuilder};
 use crate::prepare::Prepare;
 use crate::{Crate, Toolchain, Workspace};
 use failure::Error;
-use remove_dir_all::remove_dir_all;
 use std::path::PathBuf;
 use std::vec::Vec;
 
@@ -146,7 +145,7 @@ impl BuildDirectory {
     ) -> Result<R, Error> {
         let source_dir = self.source_dir();
         if source_dir.exists() {
-            remove_dir_all(&source_dir)?;
+            crate::utils::remove_dir_all(&source_dir)?;
         }
 
         let mut prepare = Prepare::new(&self.workspace, toolchain, krate, &source_dir, patches);
@@ -159,7 +158,7 @@ impl BuildDirectory {
             sandbox,
         })?;
 
-        remove_dir_all(&source_dir)?;
+        crate::utils::remove_dir_all(&source_dir)?;
         Ok(res)
     }
 
@@ -167,7 +166,7 @@ impl BuildDirectory {
     pub fn purge(&mut self) -> Result<(), Error> {
         let build_dir = self.build_dir();
         if build_dir.exists() {
-            remove_dir_all(build_dir)?;
+            crate::utils::remove_dir_all(&build_dir)?;
         }
         Ok(())
     }