+++ src/crates/git.rs
@@ -121,7 +121,7 @@ impl CrateTrait for GitRepo {
     fn purge_from_cache(&self, workspace: &Workspace) -> Result<(), Error> {
         let path = self.cached_path(workspace);
         if path.exists() {
-            remove_dir_all::remove_dir_all(&path)?;
+            crate::utils::remove_dir_all(&path)?;
         }
         Ok(())
     }