+++ src/crates/mod.rs
@@ -5,7 +5,6 @@ mod local;
 use crate::Workspace;
 use failure::Error;
 use log::info;
-use remove_dir_all::remove_dir_all;
 use std::path::Path;
 
 trait CrateTrait: std::fmt::Display {
@@ -69,7 +68,7 @@ impl Crate {
                 "crate source directory {} already exists, cleaning it up",
                 dest.display()
             );
-            remove_dir_all(dest)?;
+            crate::utils::remove_dir_all(dest)?;
         }
         self.as_trait().copy_source_to(workspace, dest)
     }