+++ src/prepare.rs
@@ -2,7 +2,6 @@ use crate::cmd::Command;
 use crate::{build::CratePatch, Crate, Toolchain, Workspace};
 use failure::{Error, Fail, ResultExt};
 use log::info;
-use std::fs::remove_file;
 use std::path::Path;
 use toml::{
     value::{Array, Table},
@@ -73,8 +72,8 @@ impl<'a> Prepare<'a> {
     fn remove_cargo_config(&self) -> Result<(), Error> {
         let path = self.source_dir.join(".cargo").join("config");
         if path.exists() {
-            remove_file(path.clone())?;
-            info!("removed {}", path.as_path().display());
+            crate::utils::remove_file(&path)?;
+            info!("removed {}", path.display());
         }
         Ok(())
     }