+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - allow workspaces by having validate_manifest now use `metadata --no-deps` instead of deprecated `read-manifest`
   therefor no longer failing on workspaces and TomlTweaker no longer removing the workspace table from `Cargo.toml`
 - `Command` now warns when it is not used.
+- remove directory/file error now mentions path
 
 ## [0.10.0] - 2020-08-08